pos-edi/services/hq-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
pydantic==2.8.2
SQLAlchemy==2.0.32
pymysql==1.1.1
httpx==0.27.0
python-dotenv==1.0.1
kafka-python==2.0.2
